BE: Started cleanup
- Deleted guest and login folder, as they are not needed
- Updated dependencies, to close Dependabot prs
- Moved all interfaces in one folder
