+++ backend/package.json
@@ -26,12 +26,12 @@
   },
   "dependencies": {
     "@nestjs/axios": "^3.0.2",
-    "@nestjs/common": "^10.0.0",
+    "@nestjs/common": "^10.3.9",
     "@nestjs/config": "^3.2.2",
     "@nestjs/core": "^10.0.0",
     "@nestjs/platform-express": "^10.0.0",
     "@ory/client": "^1.9.0",
-    "@prisma/client": "^5.14.0",
+    "@prisma/client": "^5.15.0",
     "axios": "^1.7.2",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.14.1",
@@ -47,16 +47,16 @@
     "@nestjs/testing": "^10.0.0",
     "@types/express": "^4.17.17",
     "@types/jest": "^29.5.2",
-    "@types/node": "^20.14.0",
+    "@types/node": "^20.14.2",
     "@types/supertest": "^6.0.0",
     "@typescript-eslint/eslint-plugin": "^7.11.0",
-    "@typescript-eslint/parser": "^7.11.0",
+    "@typescript-eslint/parser": "^7.13.0",
     "eslint": "^8.42.0",
     "eslint-config-prettier": "^9.0.0",
     "eslint-plugin-prettier": "^5.0.0",
     "jest": "^29.5.0",
     "prettier": "^3.3.0",
-    "prisma": "^5.14.0",
+    "prisma": "^5.15.0",
     "source-map-support": "^0.5.21",
     "supertest": "^7.0.0",
     "ts-jest": "^29.1.4",
