--- backend/src/guests/guests.controller.ts
@@ -1,4 +1,0 @@
-import { Controller } from '@nestjs/common';
-
-@Controller('guests')
-export class GuestsController {}
