+++ backend/src/interfaces/favourites.interface.ts
