+++ backend/src/interfaces/user.interface.ts
@@ -1,0 +1,21 @@
+export interface IUser {
+  userID?: string;
+  username?: string;
+  profilePicture?: null;
+  registrationDate: Date | undefined;
+  accountState: Registration;
+  expertTopics: IExpertTopics[];
+  userStatistics: object;
+  activityPoints: number;
+}
+
+export interface IExpertTopics {
+  tagname: string;
+  expertPoints: number;
+}
+
+export enum Registration {
+  registered = 'registered',
+  proUser = 'proUser',
+  admin = 'admin',
+}
