--- backend/src/login/login.controller.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { LoginController } from './login.controller';
-
-describe('LoginController', () => {
-  let controller: LoginController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [LoginController],
-    }).compile();
-
-    controller = module.get<LoginController>(LoginController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
