--- backend/src/login/login.controller.ts
@@ -1,4 +1,0 @@
-import { Controller } from '@nestjs/common';
-
-@Controller('login')
-export class LoginController {}
