+++ backend/src/requests/user-content-request/user-content-request.service.ts
@@ -32,8 +32,8 @@ import {
   IAnswer,
   IQuestion,
   IQuestionMetadata,
-} from '../questions/dto/user-content-interface';
-import { UserContentWithRating } from '../../database/user-content/user-content-interfaces';
+} from '../../interfaces/user-content.interface';
+import { UserContentWithRating } from '../../interfaces/user-content.interface';
 
 @Injectable()
 export class UserContentRequestService {
